refactor: StatsService, EndpointHitMapper extracted into a separate class; delete: EventService, will be developed in main-svc branch
