+++ stats-service/service/src/main/java/ru/practicum/EndpointHitMapper.java
@@ -1,0 +1,13 @@
+package ru.practicum;
+
+public class EndpointHitMapper {
+
+    public static EndpointHit toEndpointHit(EndpointHitDto endpointHitDto) {
+        EndpointHit endpointHit = new EndpointHit();
+        endpointHit.setApp(endpointHitDto.getApp());
+        endpointHit.setIp(endpointHitDto.getIp());
+        endpointHit.setUri(endpointHitDto.getUri());
+        endpointHit.setTimestamp(endpointHitDto.getTimestamp());
+        return endpointHit;
+    }
+}
