+++ stats-service/service/src/main/java/ru/practicum/StatsService.java
@@ -43,15 +43,4 @@ public class StatsService {
         }
         return stats;
     }
-
-    static class EndpointHitMapper {
-        public static EndpointHit toEndpointHit(EndpointHitDto endpointHitDto) {
-            EndpointHit endpointHit = new EndpointHit();
-            endpointHit.setApp(endpointHitDto.getApp());
-            endpointHit.setIp(endpointHitDto.getIp());
-            endpointHit.setUri(endpointHitDto.getUri());
-            endpointHit.setTimestamp(endpointHitDto.getTimestamp());
-            return endpointHit;
-        }
-    }
 }
